Fix size estimation for MPEG-4 transport streams.  This can reduce the number of programs that TiVo deletes upon their transfer.
TiVo requests the expected transfer size *before* it tries to actually transfer
a file, and it does not send a desired MIME type along with that request.  When
that MIME type is missing and the vCodec is "h264", tivo_compatible_video()
returned false -- which resulted in __est_size() estimating the size as if it
will transcode the video.  And, without also having this fix in
tivo_compatible_container(), that estimated size can be 2-3 times too big
because the bitrate may not reflect what's in the actual file.